+++ apps/directorate-of-equality-web/src/components/list-page/tabs/TabContent.tsx
